@@ -1,7 +1,7 @@
 import axios from 'axios';
 
 const instance = axios.create({
-    baseURL: 'http://192.168.1.219:8000'
+    baseURL: process.env.REACT_APP_API_URL || 'http://localhost:8000'
 })
 
 export default instance;
